more agressive optimizations

+++ benches/benchmarks.rs
@@ -52,11 +52,11 @@ fn downsample_stokes(c: &mut Criterion) {
                     for i in 0..iters {
                         avg_buf[idx] = payloads[i as usize].stokes_i();
                         // If we're at the end, calculate the average
-                        if idx == downsample_factor as usize - 1 {
+                        if idx == downsample_factor - 1 {
                             // Find the average into an f32 (which is lossless)
                             let mut avg = [0f32; CHANNELS];
                             for chan in 0..CHANNELS {
-                                for avg_row in avg_buf.iter().take(downsample_factor as usize) {
+                                for avg_row in avg_buf.iter().take(downsample_factor) {
                                     avg[chan] += f32::from(avg_row[chan]);
                                 }
                             }
@@ -64,7 +64,7 @@ fn downsample_stokes(c: &mut Criterion) {
                                 .for_each(|v| *v /= f32::from(downsample_factor as u16));
                         }
                         // Increment the idx
-                        idx = (idx + 1) % downsample_factor as usize;
+                        idx = (idx + 1) % downsample_factor;
                     }
                     start.elapsed()
                 })
